+++ frontend/package.json
@@ -18,33 +18,22 @@
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@headlessui/react": "^1.7.13",
-    "@nomicfoundation/hardhat-chai-matchers": "^1.0.5",
-    "@nomicfoundation/hardhat-network-helpers": "^1.0.7",
-    "@nomicfoundation/hardhat-toolbox": "^2.0.0",
-    "@nomiclabs/hardhat-ethers": "^2.2.1",
-    "@nomiclabs/hardhat-etherscan": "^3.1.3",
-    "@openzeppelin/contracts": "^4.8.0",
     "@rainbow-me/rainbowkit": "^0.8.1",
-    "@typechain/ethers-v5": "^10.2.0",
-    "@typechain/hardhat": "^6.1.5",
-    "@types/mocha": "^9.1.1",
-    "chai": "^4.3.7",
     "clsx": "^1.2.1",
     "daisyui": "^2.38.1",
     "dotenv": "^16.0.3",
     "eslint": "8.27.0",
     "eslint-config-next": "13.0.2",
     "ethers": "^5.7.2",
+    "flowbite": "^1.6.4",
+    "flowbite-react": "^0.4.2",
     "formik": "^2.2.9",
     "framer-motion": "^9.0.2",
-    "hardhat": "^2.12.4",
-    "hardhat-gas-reporter": "^1.0.9",
     "next": "13.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.7.1",
     "react-scroll": "^1.8.8",
-    "solidity-coverage": "^0.8.2",
     "ts-node": "^10.9.1",
     "typechain": "^8.1.1",
     "typescript": "^4.9.4",
