+++ frontend/tailwind.config.js
@@ -1,6 +1,9 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['./src/**/*.{js,ts,jsx,tsx}'],
+  content: [
+    './src/**/*.{js,ts,jsx,tsx}',
+    './node_modules/flowbite-react/**/*.js',
+  ],
   theme: {
     extend: {
       fontFamily: {
@@ -18,7 +21,7 @@ module.exports = {
       },
     },
   },
-  plugins: [require('daisyui')],
+  plugins: [require('daisyui', 'flowbite/plugin')],
   daisyui: {
     themes: false,
   },
